tools/requirements.txt
pyimgur==0.6.0
Pillow==10.1.0
python-dotenv==0.19.0